+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Generation of fields for CSV source, if not given.
+
 ## [1.0.0] - 2025-09-03
 
 ### Added